+++ cmd/root.go
@@ -24,13 +24,17 @@ package cmd
 import (
 	"os"
 
+	"github.com/SpyrosMoux/pwm/internal/crypto"
 	"github.com/SpyrosMoux/pwm/internal/helpers"
+	"github.com/SpyrosMoux/pwm/internal/secrets"
+	"github.com/SpyrosMoux/pwm/internal/store"
 	"github.com/spf13/cobra"
 )
 
 var (
 	storageLocation string
 	cipherKey       string
+	secretsService  *secrets.Service
 )
 
 // rootCmd represents the base command when called without any subcommands
@@ -53,7 +57,7 @@ var rootCmd = &cobra.Command{
 				os.Exit(1)
 			}
 
-			secret, err := GetSecret(name)
+			secret, err := secretsService.GetSecret(name)
 			if err != nil {
 				helpers.PrintError(err.Error())
 				os.Exit(1)
@@ -83,6 +87,19 @@ func init() {
 	homeDir, _ := os.UserHomeDir()
 	storageLocation = homeDir + "/.pwm"
 	loadCipherKey()
+
+	storer := &store.FileStore{
+		storageLocation,
+	}
+
+	crypter := &crypto.AESGCM{
+		[]byte(cipherKey),
+	}
+
+	secretsService = &secrets.Service{
+		Store:  storer,
+		Crypto: crypter,
+	}
 
 	// rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.pwm.yaml)")
 	rootCmd.PersistentFlags().StringVar(&storageLocation, "location", storageLocation, "The directory to store the "+
