+++ internal/secrets/service.go
@@ -1,0 +1,11 @@
+package secrets
+
+import (
+	"github.com/SpyrosMoux/pwm/internal/crypto"
+	"github.com/SpyrosMoux/pwm/internal/store"
+)
+
+type Service struct {
+	Store  store.Storer
+	Crypto crypto.Crypter
+}
